Fix occasional missed events
Sometimes when a file was added and then quickly renamed, only the
first change would trigger an event. This was the case because the
watched directories were registered with `EV_ONESHOT` instead of
`EV_CLEAR`.

An example of what could happen is:

1. Directory is added with `EV_ONESHOT`, and `kevent` blocks.
2. File `a.txt` is added.
3. The call to `kevent` returns, deregistering the directory due to
  `EV_ONESHOT`.
4. File `a.txt` is renamed to `b.txt`.
5. Directory is re-registered, but too late to pick up the second event.

Replacing `EV_ONESHOT` with `EV_CLEAR` fixes this, since it doesn't
deregister the directory, but instead only removes the returned events.
Registering the directory every loop should be fine, since, quoting the
man page:

> An (`ident`, `filter`, and optional `udata` value) tuple can only
> appear once in a given `kqueue`. Subsequent attempts to register the
> same tuple for a given `kqueue` will result in the replacement of the
> conditions being watched, not an addition.

That is, each successive registration should do nothing.